Move these tests to their own file.
